Fix to feedforward network
Fixed a noticeable bug in the Feedforward network that would cause
crashes or halts during backdrop. By not adding the initial input layer
to the activations array, the proper matrixes wouldn’t be formed,
causing the program to report that the vectors were not of the correct
length when subtracting the change in the weights from the current
weights.